Captured values for the attributes mentioned in #1
I also got rid of the 'processTag' function, and created it's replacement ('processAttribute').
The 'processAttribute' function has no problem dealing with value offsets, thus #5 isn't valid anymore.

I also started making the changes that are necessary to allow me to store attribute values in the attribute container (as mentioned in #3).

So far, I can only capture and store the following types:

 - EXIF_BYTE
 - EXIF_ASCII
 - EXIF_RATIONAL

Support for storing of EXIF_SHORTs and EXIF_LONGs, should be added in a commit coming in the near-future.

Signed-off-by: Gigabyte-Giant <brynden.bielefeld@hotmail.com>